[bees] Split DevTools into standalone HiveTool server
## What
Extracts the Bees devtools UI out of `packages/bees/web/` into a new `packages/bees/hivetool/` directory with its own Vite server. Introduces `packages/bees/common/` for the small shared surface (`TicketData` type + `extractPrompt`/`extractChoices` helpers).

## Why
DevTools and the Opal shell are two distinct services serving different audiences. Running them in a single Vite project couples their build, dev server, and deployment. Splitting them gives each service independent lifecycle while keeping the shared backend API contract in one canonical place.

## Changes

### New: `packages/bees/common/`
- `types.ts` — `TicketData` interface (backend API contract)
- `utils.ts` — `extractPrompt`, `extractChoices`, `Choice` (shared ticket helpers)

### New: `packages/bees/hivetool/`
- Standalone Vite project on port **5174**
- `src/ui/` — all devtools components (moved from `web/src/ui/`)
- `src/data/` — LogStore, TicketStore, StateAccess, router + Log types (moved from `web/src/data/`)
- `src/utils.ts` — devtools-only helpers (`getRelativeTime`, `parseTags`) + re-exports shared utils
- `tests/utils.test.ts` — moved from `web/tests/`

### Modified: `packages/bees/web/`
- Removed `devtools.html`, `src/ui/`, `src/data/`, `src/styles.css`, `src/constants.ts`, `src/utils.ts`
- Removed devtools entry from `vite.config.ts`
- Updated 10 source/test files to import `TicketData`/`Choice`/utils from `common/`

### Modified: `packages/bees/package.json`
- `dev` script now starts three processes: `dev:server` (Python, :3200) + `dev:web` (Opal, :5173) + `dev:hivetool` (DevTools, :5174)

## Testing
- `tsc --noEmit` passes for both `web/` and `hivetool/`
- `npm test` in `hivetool/`: 21/21 pass
- `npm test` in `web/`: 43/44 pass (1 pre-existing failure in stage-actions)
- `npm run dev -w packages/bees`: all three servers start, backend proxying works
