+++ packages/bees/common/utils.ts
@@ -4,31 +4,14 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import type { TicketData } from "./data/types.js";
+import type { TicketData } from "./types.js";
 
-export { getRelativeTime, extractPrompt, extractChoices, parseTags };
+export { extractPrompt, extractChoices };
 export type { Choice };
 
 interface Choice {
   id: string;
   text: string;
-}
-
-function getRelativeTime(isoString?: string): string {
-  if (!isoString) return "";
-  const now = new Date();
-  const past = new Date(isoString);
-  const diffMs = now.getTime() - past.getTime();
-  const diffSec = Math.floor(diffMs / 1000);
-  const diffMin = Math.floor(diffSec / 60);
-  const diffHr = Math.floor(diffMin / 60);
-  const diffDay = Math.floor(diffHr / 24);
-
-  if (diffDay > 0) return `${diffDay} day${diffDay > 1 ? "s" : ""} ago`;
-  if (diffHr > 0) return `${diffHr} hour${diffHr > 1 ? "s" : ""} ago`;
-  if (diffMin > 0) return `${diffMin} minute${diffMin > 1 ? "s" : ""} ago`;
-  if (diffSec > 0) return `${diffSec} second${diffSec > 1 ? "s" : ""} ago`;
-  return "just now";
 }
 
 function extractPrompt(ticket: TicketData): string {
@@ -57,10 +40,4 @@ function extractChoices(ticket: TicketData): Choice[] {
     const texts = parts.filter((p) => p.text).map((p) => p.text);
     return { id, text: texts.join("\n") || id };
   });
-}
-function parseTags(text: string): string[] {
-  return text
-    .split(",")
-    .map((t) => t.trim())
-    .filter(Boolean);
 }
