+++ packages/bees/hivetool/package.json
@@ -1,26 +1,20 @@
 {
-  "name": "bees-web",
+  "name": "bees-hivetool",
   "private": true,
   "type": "module",
   "scripts": {
     "dev": "vite",
     "build:test": "tsc -p tsconfig.test.json",
-    "test": "npm run build:test && node --test './dist/tsc/web/tests/**/*.test.js'",
-    "lint": "eslint --max-warnings=0 'src/**/*.ts'"
+    "test": "npm run build:test && node --test './dist/tsc/hivetool/tests/**/*.test.js'"
   },
   "dependencies": {
     "@lit-labs/signals": "^0.2.0",
-    "@lit/context": "^1.1.6",
     "@types/markdown-it": "^14.1.2",
     "lit": "^3.2.0",
     "markdown-it": "^14.1.1",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
     "signal-polyfill": "^0.2.2"
   },
   "devDependencies": {
-    "@types/react": "^19.2.14",
-    "@types/react-dom": "^19.2.3",
     "typescript": "^5.7.0",
     "vite": "^6.4.2"
   }
