+++ packages/bees/hivetool/src/data/log-store.ts
