+++ packages/bees/hivetool/src/data/types.ts
@@ -4,41 +4,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export interface TicketData {
-  id: string;
-  objective: string;
-  status: string;
-  created_at?: string;
-  completed_at?: string;
-  turns?: number;
-  thoughts?: number;
-  outcome?: string;
-  error?: string;
-  assignee?: string;
-  suspend_event?: Record<string, unknown>;
-  depends_on?: string[];
-  events_log?: Array<Record<string, unknown>>;
-  tags?: string[];
-  functions?: string[];
-  skills?: string[];
-  title?: string;
-  playbook_id?: string;
-  playbook_run_id?: string;
-  parent_ticket_id?: string;
-  creator_ticket_id?: string;
-  slug?: string;
-  kind?: string;
-  signal_type?: string;
-  context?: string;
-  delivered_to?: string[];
-  chat_history?: Array<{ role: string; text: string }>;
-  model?: string;
-  watch_events?: Array<{ type: string }>;
-  outcome_content?: Record<string, unknown>;
-  files?: Array<{ path: string; mimeType: string; localPath: string }>;
-}
-
-
+// Re-export the shared backend API contract.
+export type { TicketData } from "../../../common/types.js";
 
 // ---------------------------------------------------------------------------
 // Log file types (from EvalCollector output in packages/bees/hive/logs)
