+++ packages/bees/hivetool/src/ui/log-detail.styles.ts
