+++ packages/bees/hivetool/src/ui/log-detail.ts
