+++ packages/bees/hivetool/src/ui/truncated-text.ts
