+++ packages/bees/hivetool/src/utils.ts
@@ -1,0 +1,37 @@
+/**
+ * @license
+ * Copyright 2026 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+
+
+// Re-export shared helpers from common.
+export { extractPrompt, extractChoices } from "../../common/utils.js";
+export type { Choice } from "../../common/utils.js";
+
+export { getRelativeTime, parseTags };
+
+function getRelativeTime(isoString?: string): string {
+  if (!isoString) return "";
+  const now = new Date();
+  const past = new Date(isoString);
+  const diffMs = now.getTime() - past.getTime();
+  const diffSec = Math.floor(diffMs / 1000);
+  const diffMin = Math.floor(diffSec / 60);
+  const diffHr = Math.floor(diffMin / 60);
+  const diffDay = Math.floor(diffHr / 24);
+
+  if (diffDay > 0) return `${diffDay} day${diffDay > 1 ? "s" : ""} ago`;
+  if (diffHr > 0) return `${diffHr} hour${diffHr > 1 ? "s" : ""} ago`;
+  if (diffMin > 0) return `${diffMin} minute${diffMin > 1 ? "s" : ""} ago`;
+  if (diffSec > 0) return `${diffSec} second${diffSec > 1 ? "s" : ""} ago`;
+  return "just now";
+}
+
+function parseTags(text: string): string[] {
+  return text
+    .split(",")
+    .map((t) => t.trim())
+    .filter(Boolean);
+}
