+++ packages/bees/package.json
@@ -10,10 +10,11 @@
     "ticket:drain": ".venv/bin/python -m bees.drain",
     "ticket:respond": ".venv/bin/python -m bees.respond",
     "playbook:run": ".venv/bin/python -m bees.run_playbook",
-    "dev": "npm run dev:server & npm run dev:web",
+    "dev": "npm run dev:server & npm run dev:web & npm run dev:hivetool",
     "dev:clean": "rm -rf hive && npm run dev",
     "dev:server": ".venv/bin/python -m bees.server",
     "dev:web": "cd web && npm run dev",
+    "dev:hivetool": "cd hivetool && npm run dev",
     "test:python": ".venv/bin/python -m pytest tests/ -v"
   },
   "wireit": {
